Update script.js
fixed bug when taking one month away from 0 months and removed unnecessary "if" for when months == 12.  

The number of months used to go to -1 and as a result the months output would animate to infinity. Now the nr. of months goes from 0 to 11 and the nr. of years goes down by 1 aswell.